fixed syntax errors

+++ src/apps/X2DemoMachine/states/X2DemoState.h
@@ -62,7 +62,7 @@ public:
     double rateLimit;
     double refPos1;
     double refPos2;
-    double refPosPeriod;
+    int refPosPeriod;
     GroupController<PDController<double>, X2_NUM_JOINTS> jointControllers;
     Eigen::VectorXd debugTorques;
     Eigen::VectorXd frictionCompensationTorques;
